requirements.txt
tensorflow-cpu
pillow
tk
flask
requests
numpy
pylint
jupyter
